TxButton fix
